Massive refactoring to add namespaces
Namespaces are added to the core files and tests and this drops support
for PHP 5.2

This may have broken some features of the plugin. Need to test and
rework.

This will break the current add-ons. They need to be reworked.

During activation, PHP version is checked and if it is less than PHP 5.3
then the plugin deactivates itself with a warning.

Fix #45